backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
fpdf==1.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
gunicorn==21.2.0
flask_cors
python-dotenv==1.0.1
pytz
google-generativeai
pymongo[srv]
